Set up expanded state for StudentCard
This commit adds `expanded` boolean state to the `<StudenCard />`
component. When this state is true, the card will display all of the
grades. When false, greades will not be shown.

Future commits will implement a button to toggel the state between
true and false.

For the moment, ** there are no user facing changes** in this commit.